Break doGitForMe.js up into multiple modules and add --hs to wait for --grunt.
